fix(mongodb): increment id counters with an int64 operand
Copilot (suppressed comment on the re-review of 4ca6ffb) pointed out that
the upserting $inc in nextSequence used an untyped Go literal, which the
driver encodes as BSON int32, so a fresh counter backing int64 IDs was
created as a 32-bit integer. Use an explicit int64(1) so the counter type
matches the ID fields from the first allocation.

+++ internal/storage/mongodb/sequence.go
@@ -13,14 +13,16 @@ import (
 // the new value, creating the counter on first use. The update must return
 // the document *after* the increment: with the driver default (before) the
 // first two callers on a fresh database both receive 1 and the second insert
-// fails with a duplicate _id.
+// fails with a duplicate _id. The increment operand is an explicit int64 so
+// that the upsert creates the counter as a 64-bit integer matching the ID
+// fields, rather than the int32 an untyped literal would encode to.
 func nextSequence(ctx context.Context, counters *mongo.Collection, key string) (int64, error) {
 	var doc struct {
 		Value int64 `bson:"value"`
 	}
 	err := counters.FindOneAndUpdate(ctx,
 		bson.M{"_id": key},
-		bson.M{"$inc": bson.M{"value": 1}},
+		bson.M{"$inc": bson.M{"value": int64(1)}},
 		options.FindOneAndUpdate().SetUpsert(true).SetReturnDocument(options.After),
 	).Decode(&doc)
 	if err != nil {
